requirements.txt
matplotlib==3.7.2
mplcursors==0.5.2
numpy==1.23.2
pandas==2.0.3
Pillow==10.4.0